requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
google-cloud-aiplatform==1.71.0
supabase==2.10.0
python-dotenv==1.0.1
